Code-Evaluator-V2/requirements.txt
huggingface_hub>=0.24.0
pyyaml>=6.0
ruff>=0.1.0
